Rename diff-filtering-utils.ts to diff-informed-analysis-utils.ts
