+++ src/diff-informed-analysis-utils.ts
